backend/python/requirements.txt
# Python requirements
surya-ocr>=0.4.0
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
numpy>=1.21.0
opencv-python>=4.5.0
transformers>=4.20.0
datasets>=2.0.0
